Create task to database

+++ app/src/main/java/com/example/todo/data/network/entity/NewTask.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class NewTask(
     val title: String,
     val description: String,
-    val deadline: Long?,
+    val deadline: Int?,
     @SerializedName("category_id") val category: Int,
     @SerializedName("priority_id") val priority: Int,
     val done:Int = 0,
